+++ src/main/java/com/moonsworth/lunar/client/feature/type/waypoints/WaypointRenderer.java
@@ -3,7 +3,7 @@ package com.moonsworth.lunar.client.feature.type.waypoints;
 import org.lwjgl.opengl.GL11;
 
 public class WaypointRenderer
-extends WapointRendererBase {
+    extends WapointRendererBase {
     public void lIlIlIlIlIIlIIlIIllIIIIIl(float f, float f2, float f3, int n, int n2) {
         float f4 = this.IlllIIIIIIlllIlIIlllIlIIl == 100021 ? -1.0f : 1.0f;
         float f5 = (float)Math.PI * 2 / (float)n;
@@ -11,7 +11,7 @@ extends WapointRendererBase {
         float f7 = f3 / (float)n2;
         float f8 = (f - f2) / f3;
         if (this.lIlIlIlIlIIlIIlIIllIIIIIl == 100010) {
-            GL11.glBegin((int)0);
+            GL11.glBegin(0);
             for (int i = 0; i < n; ++i) {
                 float f9 = this.IlllIIIIIIlllIlIIlllIlIIl((float)i * f5);
                 float f10 = this.lIlIlIlIlIIlIIlIIllIIIIIl((float)i * f5);
@@ -19,7 +19,7 @@ extends WapointRendererBase {
                 float f11 = 0.0f;
                 float f12 = f;
                 for (int j = 0; j <= n2; ++j) {
-                    GL11.glVertex3f((float)(f9 * f12), (float)(f10 * f12), (float)f11);
+                    GL11.glVertex3f(f9 * f12, f10 * f12, f11);
                     f11 += f7;
                     f12 += f6;
                 }
@@ -34,7 +34,7 @@ extends WapointRendererBase {
                 float f17 = f;
                 for (int i = 0; i < n2; ++i) {
                     float f18 = 0.0f;
-                    GL11.glBegin((int)8);
+                    GL11.glBegin(8);
                     for (int j = 0; j <= n; ++j) {
                         float f19;
                         float f20;
@@ -48,17 +48,17 @@ extends WapointRendererBase {
                         if (f4 == 1.0f) {
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f20 * f4, f19 * f4, f8 * f4);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f18, f15);
-                            GL11.glVertex3f((float)(f20 * f17), (float)(f19 * f17), (float)f16);
+                            GL11.glVertex3f(f20 * f17, f19 * f17, f16);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f20 * f4, f19 * f4, f8 * f4);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f18, f15 + f14);
-                            GL11.glVertex3f((float)(f20 * (f17 + f6)), (float)(f19 * (f17 + f6)), (float)(f16 + f7));
+                            GL11.glVertex3f(f20 * (f17 + f6), f19 * (f17 + f6), f16 + f7);
                         } else {
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f20 * f4, f19 * f4, f8 * f4);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f18, f15);
-                            GL11.glVertex3f((float)(f20 * f17), (float)(f19 * f17), (float)f16);
+                            GL11.glVertex3f(f20 * f17, f19 * f17, f16);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f20 * f4, f19 * f4, f8 * f4);
                             this.lIlIlIlIlIIlIIlIIllIIIIIl(f18, f15 + f14);
-                            GL11.glVertex3f((float)(f20 * (f17 + f6)), (float)(f19 * (f17 + f6)), (float)(f16 + f7));
+                            GL11.glVertex3f(f20 * (f17 + f6), f19 * (f17 + f6), f16 + f7);
                         }
                         f18 += f13;
                     }
@@ -76,45 +76,44 @@ extends WapointRendererBase {
                 float f23 = 0.0f;
                 float f24 = f;
                 for (int i = 0; i <= n2; ++i) {
-                    GL11.glBegin((int)2);
+                    GL11.glBegin(2);
                     for (n3 = 0; n3 < n; ++n3) {
                         f22 = this.IlllIIIIIIlllIlIIlllIlIIl((float)n3 * f5);
                         f21 = this.lIlIlIlIlIIlIIlIIllIIIIIl((float)n3 * f5);
                         this.lIlIlIlIlIIlIIlIIllIIIIIl(f22 * f4, f21 * f4, f8 * f4);
-                        GL11.glVertex3f((float)(f22 * f24), (float)(f21 * f24), (float)f23);
+                        GL11.glVertex3f(f22 * f24, f21 * f24, f23);
                     }
                     GL11.glEnd();
                     f23 += f7;
                     f24 += f6;
                 }
             } else if ((double)f != 0.0) {
-                GL11.glBegin((int)2);
+                GL11.glBegin(2);
                 for (n3 = 0; n3 < n; ++n3) {
                     f22 = this.IlllIIIIIIlllIlIIlllIlIIl((float)n3 * f5);
                     f21 = this.lIlIlIlIlIIlIIlIIllIIIIIl((float)n3 * f5);
                     this.lIlIlIlIlIIlIIlIIllIIIIIl(f22 * f4, f21 * f4, f8 * f4);
-                    GL11.glVertex3f((float)(f22 * f), (float)(f21 * f), (float)0.0f);
+                    GL11.glVertex3f(f22 * f, f21 * f, 0.0f);
                 }
                 GL11.glEnd();
-                GL11.glBegin((int)2);
+                GL11.glBegin(2);
                 for (n3 = 0; n3 < n; ++n3) {
                     f22 = this.IlllIIIIIIlllIlIIlllIlIIl((float)n3 * f5);
                     f21 = this.lIlIlIlIlIIlIIlIIllIIIIIl((float)n3 * f5);
                     this.lIlIlIlIlIIlIIlIIllIIIIIl(f22 * f4, f21 * f4, f8 * f4);
-                    GL11.glVertex3f((float)(f22 * f2), (float)(f21 * f2), (float)f3);
+                    GL11.glVertex3f(f22 * f2, f21 * f2, f3);
                 }
                 GL11.glEnd();
             }
-            GL11.glBegin((int)1);
+            GL11.glBegin(1);
             for (n3 = 0; n3 < n; ++n3) {
                 f22 = this.IlllIIIIIIlllIlIIlllIlIIl((float)n3 * f5);
                 f21 = this.lIlIlIlIlIIlIIlIIllIIIIIl((float)n3 * f5);
                 this.lIlIlIlIlIIlIIlIIllIIIIIl(f22 * f4, f21 * f4, f8 * f4);
-                GL11.glVertex3f((float)(f22 * f), (float)(f21 * f), (float)0.0f);
-                GL11.glVertex3f((float)(f22 * f2), (float)(f21 * f2), (float)f3);
+                GL11.glVertex3f(f22 * f, f21 * f, 0.0f);
+                GL11.glVertex3f(f22 * f2, f21 * f2, f3);
             }
             GL11.glEnd();
         }
     }
 }
-
